pyproject.toml
[project]
name = "ml-deepglobe"
version = "0.1.0"
description = "Semantic Segmentation of Satellite Imagery using Deep Learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py==2.3.1",
    "affine==2.4.0",
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "click-plugins==1.1.1.2",
    "cligj==0.7.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "filelock==3.13.1",
    "fonttools==4.60.1",
    "fsspec==2024.6.1",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "grpcio==1.75.1",
    "hf-xet==1.1.10",
    "huggingface-hub==0.35.3",
    "idna==3.10",
    "ipykernel>=7.1.0",
    "jinja2==3.1.3",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "lightning-utilities==0.15.2",
    "markdown==3.9",
    "markupsafe==2.1.5",
    "matplotlib==3.10.6",
    "mpmath==1.3.0",
    "networkx==3.3",
    "numpy==2.1.2",
    "opencv-python-headless==4.12.0.88",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==11.0.0",
    "platformdirs==4.4.0",
    "protobuf==6.32.1",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "rasterio==1.4.3",
    "requests==2.32.5",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "segmentation-models-pytorch==0.5.0",
    "sentry-sdk==2.39.0",
    "simsimd==6.5.3",
    "six==1.17.0",
    "smmap==5.0.2",
    "stringzilla==4.0.15",
    "sympy==1.13.1",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "threadpoolctl==3.6.0",
    "timm==1.0.20",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchmetrics==1.8.2",
    "torchvision==0.20.1",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wandb==0.22.1",
    "werkzeug==3.1.3",
]
[project.optional-dependencies]
cuda = [
    "nvidia-cublas-cu11==11.11.3.6",
    "nvidia-cuda-cupti-cu11==11.8.87",
    "nvidia-cuda-nvrtc-cu11==11.8.89",
    "nvidia-cuda-runtime-cu11==11.8.89",
    "nvidia-cudnn-cu11==9.1.0.70",
    "nvidia-cufft-cu11==10.9.0.58",
    "nvidia-curand-cu11==10.3.0.86",
    "nvidia-cusolver-cu11==11.4.1.48",
    "nvidia-cusparse-cu11==11.7.5.86",
    "nvidia-nccl-cu11==2.21.5",
    "nvidia-nvtx-cu11==11.8.86",
]

